fix(feed): tx detail multi-swap shows concrete token brand per leg
The Activity feed detail screen for a multi-leg 7702 Dolares -> Pesos
swap listed each leg as "0.91 Dolares" / "1.04 Dolares" / "1.04 Dolares"
so the user could not tell which stablecoin brand each portion actually
came from. The success screen already renders this breakdown correctly
via an inline TokenAmountWithBrand component (per leg: "0.91 Tether USD",
"1.04 USD Coin", "1.04 Dolar Mento").

Extract TokenAmountWithBrand from TransactionSuccessScreen into a shared
component at src/components/TokenAmountWithBrand.tsx, and use it in
SwapContent for the per-leg rows when isMultiLegSwap. The detail screen
now matches the success screen exactly: aggregate stays labelled
"Dolares" on the header row, per-leg breakdown shows the concrete brand.

No behavior change for single-leg swaps or non-dollar tokens.
